requirements.txt
jinja2>=3.1
requests>=2.28
pytest>=7.0
anthropic>=0.40
python-dotenv>=1.0
beautifulsoup4>=4.12
